Se empezó el desarrollo de las habilidades

+++ Juego_Practica/src/Models/Character.java
@@ -10,7 +10,7 @@ public class Character {
     private double inte;
     private double hp;
     private double armor;
-    private double mana;
+    private int mana;
     private ArrayList<Hability> habilities;
 
     public Character(String id, String name, String charClass, int level) {
@@ -78,7 +78,6 @@ public class Character {
     public void setStr(double str) {
         this.str = str;
     }
-
     public double getAgi() {
         agi=5;
         switch (getCharClass()){
@@ -151,7 +150,7 @@ public class Character {
         this.armor = armor;
     }
 
-    public double getMana() {
+    public int getMana() {
         mana=100;
         switch (getCharClass()){
             case "Warrior":
@@ -165,7 +164,7 @@ public class Character {
         }
     }
 
-    public void setMana(double mana) {
+    public void setMana(int mana) {
         this.mana = mana;
     }
